+++ src/Monad.fs
@@ -2,18 +2,20 @@ namespace MF.Monad
 
 [<RequireQualifiedAccess>]
 module Writer =
-    type Writer<'a> = Writer of ('a * string)
+    open MF.ConsoleApplication
 
-    type Kleisli<'a, 'b, 'c> = ('a -> Writer<'b>) -> ('b -> Writer<'c>) -> ('a -> Writer<'c>)
-    let (>=>): Kleisli<'a, 'b, 'c> =
+    type Writer<'A> = Writer of ('A * string)
+
+    type Kleisli<'A, 'B, 'C> = ('A -> Writer<'B>) -> ('B -> Writer<'C>) -> ('A -> Writer<'C>)
+    let (>=>): Kleisli<'A, 'B, 'C> =
         fun m1 m2 a ->
             let (Writer (b, s1)) = m1 a
             let (Writer (c, s2)) = m2 b
 
             Writer (c, s1 + s2)
 
-    type Return<'a> = 'a -> Writer<'a>
-    let retn: Return<'a> = fun a -> Writer (a, "")
+    type Return<'A> = 'A -> Writer<'A>
+    let retn: Return<'A> = fun a -> Writer (a, "")
 
     type UpCase = string -> Writer<string>
     let upCase: UpCase = fun s -> Writer(s.ToUpper(), "upCase ")
@@ -24,7 +26,7 @@ module Writer =
     type Process = string -> Writer<string list>
     let runProcess: Process = upCase >=> toWords
 
-    let example ((input, output): MF.ConsoleApplication.IO) =
+    let example = Execute <| fun (input, output) ->
         let (Writer (words, log)) =
             "Hello World"
             |> runProcess
@@ -32,4 +34,4 @@ module Writer =
         output.Messages " - " words
         output.Message <| sprintf "<c:dark-yellow>// %s</c>" log
 
-        ()
+        ExitCode.Success